addressing PR #62, part 3

+++ source/nanoFramework.Hardware.Esp32/TouchPad/TouchPadConfig.cs
@@ -10,13 +10,18 @@ namespace nanoFramework.Hardware.Esp32.TouchPad
 	/// </summary>
 	public class TouchPadConfig : TouchPadBaseConfig
 	{
-		//the threshold to trigger interrupt when the pad is touched
+		/// <summary>
+		/// The threshold to trigger interrupt when the pad is touched
+		/// </summary>
 		public float InterruptThresholdValue { get; set; }
 
+		/// <summary>
+		/// Default constructor
+		/// </summary>
 		public TouchPadConfig() : base()
 		{
-			// Use 2 / 3 of read value as the threshold to trigger interrupt when the pad is touched.
-			InterruptThresholdValue = 2 / 3;
+			// By default, use 2/3 of read value as the threshold to trigger interrupt when the pad is touched.
+			InterruptThresholdValue = (float) 2 / 3;
 		}
 	}
 }
